+++ kamehouse-shell/bin/kamehouse/docker/docker-container/docker-init-kamehouse-folder-to-defaults.sh
@@ -14,16 +14,16 @@ mainProcess() {
   mkdir -p ${HOME}/.kamehouse/config/
   cp -v -f ${HOME}/git/kamehouse/docker/config/kamehouse.cfg ${HOME}/.kamehouse/config/
 
-  log.info "Reinit .kamehouse-secrets.cfg"
-  mkdir -p ${HOME}/.kamehouse/config/keys/
-  cp -v -f ${HOME}/git/kamehouse/docker/keys/.kamehouse-secrets.cfg ${HOME}/.kamehouse/config/keys/
-
   log.info "Reinit keys"
   mkdir -p ${HOME}/.kamehouse/config/keys
   cp -v -f ${HOME}/git/kamehouse/docker/keys/.*.pwd.enc ${HOME}/.kamehouse/config/keys
   cp -v -f ${HOME}/git/kamehouse/docker/keys/integration-test-cred.enc ${HOME}/.kamehouse/config/keys
+  cp -v -f ${HOME}/git/kamehouse/docker/keys/.kamehouse-secrets.cfg.enc ${HOME}/.kamehouse/config/keys/
   cp -v -f ${HOME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pkcs12 ${HOME}/.kamehouse/config/keys/kamehouse.pkcs12
   cp -v -f ${HOME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.crt ${HOME}/.kamehouse/config/keys/kamehouse.crt
+  cp -v -f ${HOME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.key ${HOME}/.kamehouse/config/keys/kamehouse.key
+  cp -v -f ${HOME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pub ${HOME}/.kamehouse/config/keys/kamehouse.pub
+  cp -v -f ${HOME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/secrets.key.enc ${HOME}/.kamehouse/config/keys/kamehouse-secrets.key.enc
 
   log.info "Reinit mariadb dump data"
   mkdir -p ${HOME}/.kamehouse/config/mariadb/dump/old